refactor(sybra): split service wiring fan-out
