+++ internal/sybra/services.go
@@ -1,104 +1,26 @@
 package sybra
 
 import (
-	"context"
 	"maps"
 
 	"github.com/Automaat/sybra/internal/httpapi"
-	"github.com/Automaat/sybra/internal/umbrella"
 )
 
 // wireServices populates the Wails-bound service structs that were pre-allocated
 // in NewApp(). Must be called after all dependencies are initialized.
 func (a *App) wireServices(emit func(string, any)) {
-	a.reviewer.workflowEngine = a.workflowEngine
-	a.reviewSvc.reviewer = a.reviewer
-	a.reviewSvc.tasks = a.tasks
-	a.taskSvc.tasks = a.tasks
-	a.taskSvc.agents = a.agents
-	a.taskSvc.workflowEngine = a.workflowEngine
-	a.taskSvc.worktrees = a.worktrees
-	a.taskSvc.sandboxes = a.sandboxes
-	a.taskSvc.wg = &a.wg
-	a.taskSvc.logger = a.logger
-	a.taskSvc.audit = a.audit
-	a.taskSvc.cfg = a.cfg
-	// Expand a manually-added ☂️ umbrella issue into a gated child DAG instead
-	// of a flat task. Wired unconditionally; enrichFromIssue gates the call on
-	// cfg.Umbrella.Enabled so a config reload toggles it without re-wiring.
-	// Mirrors initIssuesFetcher's poll-loop expander (same Expand entry point).
-	a.taskSvc.umbrellaExpand = func(issueURL string) (umbrella.Result, error) {
-		return umbrella.Expand(context.Background(), a.tasks, umbrella.FallbackPlannerRunner(a.cfg.Umbrella.Model, a.providerHealth), issueURL)
-	}
-	a.planSvc.engine = a.workflowEngine
-	a.planSvc.tasks = a.tasks
-	a.planSvc.agents = a.agents
-	a.agentSvc.agents = a.agents
-	a.agentSvc.logger = a.logger
-	a.agentSvc.tasks = a.tasks
-	a.agentSvc.cfg = a.cfg
-	a.agentSvc.logsDir = a.logDir
-	a.agentSvc.worktrees = a.worktrees
-	a.orchSvc.agents = a.agents
-	a.orchSvc.audit = a.audit
-	a.orchSvc.logger = a.logger
-	a.orchSvc.emit = emit
-	a.agentOrch.sandboxes = a.sandboxes
-	a.projectSvc.projects = a.projects
-	a.projectSvc.worktrees = a.worktrees
-	a.projectSvc.logger = a.logger
-	a.projectSvc.notifier = a.notifier
-	a.projectSvc.bgops = a.bgops
-	a.projectSvc.wg = &a.wg
-	a.agentOrch.bgops = a.bgops
-	a.loopAgentSvc.store = a.loopAgents
-	a.loopAgentSvc.sched = a.loopSched
-	a.loopAgentSvc.auditDir = a.auditDir
-	a.loopAgentSvc.logger = a.logger
+	a.wireReviewServices()
+	a.wireTaskService()
+	a.wirePlanningService()
+	a.wireAgentServices(emit)
+	a.wireProjectServices()
+	a.wireLoopAgentService()
 	a.wireConfigService()
-	a.intgSvc.tasks = a.tasks
-	a.intgSvc.projects = a.projects
-	a.intgSvc.agents = a.agents
-	a.intgSvc.worktrees = a.worktrees
-	a.intgSvc.audit = a.audit
-	a.intgSvc.cfg = a.cfg
-	a.intgSvc.logger = a.logger
-	a.intgSvc.todoistHandler = a.todoistHandler
-	a.intgSvc.renovateHandler = a.renovateHandler
-	a.intgSvc.workflowEngine = a.workflowEngine
-	a.intgSvc.providerHealth = a.providerHealth
-	a.intgSvc.saveConfig = func() error { return a.cfg.Save() }
+	a.wireIntegrationService()
 	a.wireStatsService()
-	a.workflowSvc.engine = a.workflowEngine
-	a.workflowSvc.store = a.workflowStore
-	a.browserSvc.open = a.openBrowser
-
-	// Build completion handlers last — by this point every dependency they
-	// read is wired up, and the manager's construction-time callback delegates
-	// to a.agentCompletion once it is populated here.
-	a.agentCompletion = a.newAgentCompletionHandler(emit)
-	if a.workflowEngine != nil {
-		a.workflowEngine.SetOnComplete(a.agentCompletion.OnWorkflowComplete)
-	}
-}
-
-func (a *App) wireConfigService() {
-	a.configSvc.cfg = a.cfg
-	a.configSvc.logLevel = a.logLevel
-	a.configSvc.notifier = a.notifier
-	a.configSvc.agents = a.agents
-	a.configSvc.limits = a.limits
-	a.configSvc.logger = a.logger
-	a.configSvc.policy = a.limitPolicy
-	a.configSvc.reloadHook = a.reloadTodoist
-}
-
-func (a *App) wireStatsService() {
-	a.statsSvc.stats = a.stats
-	a.statsSvc.limits = a.limits
-	a.statsSvc.projects = a.projects
-	a.statsSvc.tasks = a.tasks
-	a.statsSvc.policy = a.limitPolicy
+	a.wireWorkflowService()
+	a.wireBrowserService()
+	a.wireCompletionHandlers(emit)
 }
 
 // ServiceRegistry returns the named service instances for HTTP dispatch.
