+++ internal/sybra/services_wire_browser.go
@@ -1,0 +1,5 @@
+package sybra
+
+func (a *App) wireBrowserService() {
+	a.browserSvc.open = a.openBrowser
+}
