+++ internal/sybra/services_wire_planning.go
@@ -1,0 +1,7 @@
+package sybra
+
+func (a *App) wirePlanningService() {
+	a.planSvc.engine = a.workflowEngine
+	a.planSvc.tasks = a.tasks
+	a.planSvc.agents = a.agents
+}
