+++ internal/sybra/services_wire_reviews.go
@@ -1,0 +1,7 @@
+package sybra
+
+func (a *App) wireReviewServices() {
+	a.reviewer.workflowEngine = a.workflowEngine
+	a.reviewSvc.reviewer = a.reviewer
+	a.reviewSvc.tasks = a.tasks
+}
