+++ internal/sybra/services_wire_stats.go
@@ -1,0 +1,9 @@
+package sybra
+
+func (a *App) wireStatsService() {
+	a.statsSvc.stats = a.stats
+	a.statsSvc.limits = a.limits
+	a.statsSvc.projects = a.projects
+	a.statsSvc.tasks = a.tasks
+	a.statsSvc.policy = a.limitPolicy
+}
