+++ internal/sybra/services_wire_workflow.go
@@ -1,0 +1,6 @@
+package sybra
+
+func (a *App) wireWorkflowService() {
+	a.workflowSvc.engine = a.workflowEngine
+	a.workflowSvc.store = a.workflowStore
+}
